updated dependencies improved io-utils improved srec-utils improved gui-utils

+++ Utils/Gui/gui-utils/gui-utils/src/main/java/com/utils/gui/factories/BasicControlsFactory.java
@@ -44,7 +44,7 @@ public interface BasicControlsFactory {
 
 	@ApiMethod
 	TextField createNumberOnlyTextField(
-			int value);
+			double value);
 
 	@ApiMethod
 	TextField createReadOnlyTextField(
